third_party/blink/web_tests/wpt_internal/display-lock/css-content-visibility/accessibility/content-visibility-accessibility-009.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<link rel="author" title="Joey Arhar" href="mailto:jarhar@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<input id=inputid aria-labelledby="hiddenuntilfound-label">
<div hidden=until-found>
  <span id="hiddenuntilfound-label">hiddenuntilfound label text</span>
</div>

<script>
test(() => {
  const axInput = accessibilityController.accessibleElementById('inputid');
  assert_equals(axInput.name, 'hiddenuntilfound label text');
}, `aria labels inside hidden=until-found subtrees should still work.`);
</script>
